refactor(scheduler): extract snapshot RPC handlers to scheduler_snapshot_handlers.py
Phase C.2 of the scheduler decomposition extraction port.

Move 11 snapshot RPC handler methods and pending-restore registry
helpers (~1078 lines, the entire former S8 block) out of monolithic
scheduler.py into a new fork-only file
python/sglang/srt/managers/scheduler_snapshot_handlers.py as
module-level functions taking a scheduler argument:

  _find_request_by_rid, _load_snapshot_for_pending_restore,
  _stage_pending_restore, _clear_pending_restore,
  _maybe_hydrate_from_pending_restore, handle_save_snapshot,
  handle_list_snapshots, handle_get_snapshot_info,
  handle_restore_snapshot, handle_delete_snapshot,
  handle_mamba_state_eviction.

Delegator pattern uses class-attribute binding for concision with 11
methods: scheduler.py has

    _stage_pending_restore = _snapshot_handlers._stage_pending_restore
    handle_save_snapshot = _snapshot_handlers.handle_save_snapshot
    ...

Each assignment turns the module function into a Scheduler method when
accessed via an instance (Python passes self as the first arg
automatically). Test surface preserved: test_pending_restore_registry.py
calls scheduler._stage_pending_restore(...) and continues to work
without modification. Phase C.1 used 1-line def delegators for 2
methods; with 11 methods, class-attribute binding is significantly more
concise without sacrificing semantics.

Per §7 R5 re-scoping, the Scheduler continues to own all snapshot/tier
attributes; this is logic relocation only, no attribute-ownership
migration. ~50+ cross-subsystem call sites elsewhere in scheduler.py
remain UNCHANGED.

Module logger bound to "sglang.srt.managers.scheduler" so log records
continue to appear under the Scheduler logger name post-extraction.

Marker policy: header-only on the new file per snapshot.py precedent
for fully fork-authored files. The scheduler.py import + delegators
are wrapped in clean BEGIN/END ENGRAM blocks.

Semantics-preserving — no behavior change. Block balance 16/15 (the
pre-existing L29 orphan END / L30 unclosed BEGIN slated for Phase F;
Phase C.2 net change was +1 import block / +1 delegator block / -1
deleted S8 wrapper = +1/+1).

py_compile passes on both touched files. Test method-call surface
preserved verbatim.
